Add hack for min-content body width.
IE and older browsers in general do not support the min-content sizing property, which sizes elements at their smallest possible size. Setting the container of said elements to display: table and setting a width of 1% simulates this property. Although it is not semantic, it was the only method that worked with only CSS changes and without altering the HTML or adding new JavaScript.

Flexbox was removed and replaced with a version of the relative positioning method for vertical centering. Because the container this is done on is sized via viewport units, the regular 50%, -50% process does not work. Knowing that the container is 100vh, positioning it 50vh from the top centers the child elements from their tops. Translating across the y-axis -32.5% then shifts the children approximately towards the center (slightly offset with more space above). Removing the flexbox dependency also extends support to IE 9.

A pollyfill for requestAnimationFrame was also added as a back up for the smooth-scroll pollyfill in an IE 9 environment.