Remove flexbox and add support for IE 9.
Flexbox broke the display: table hack that simulates the min-content property, so vertical centering on the banner required a different method. Using viewport units in combination with percentages and calc allows for approximate centering via padding-top. For this to truly work, however, a pollyfill is required for Mobile Safari, which has been added.