Adjust legibility and complexity of design
The initial redesign, while providing a foundation, had some issues with
legibility and consistency across devices. These changes make text a
stronger weight and a darker color for easier reading, and also unify
the layout for all devices.

Some features are still missing or need improvement, such as links in
the header introduction and the project/work experience showcases,
but for now the focus is on site usability and accessibility.