Switch to Next.js and Chakra UI
Please excuse the massive number of changes in this commit. This
conversion began as a fun experiment and evolved into a serious
refactor. Because it was never originally planned for any
of the changes to be published, they were unfortunately not properly
tracked from the beginning. Lesson learned: always track changes,
regardless of the end intent.